requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
Authlib==1.6.6
beautifulsoup4==4.14.3
blinker==1.9.0
brotli==1.2.0
certifi==2025.11.12
cffi==1.16.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.13
cozepy==0.20.0
cryptography==45.0.7
Cython==3.2.2
dataclass-wizard==0.39.1
ddddocr==1.5.6
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
Faker==38.2.0
flatbuffers==25.9.23
fsspec==2025.12.0
greenlet==3.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==2.3.0
intel-cmplr-lib-ur==2025.3.1
intel-openmp==2025.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaitaistruct==0.11
Levenshtein==0.27.3
lxml==6.0.2
MarkupSafe==3.0.3
mkl==2025.3.0
modin==0.37.1
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
Naked==0.1.32
Nuitka==2.8.9
numpy==2.4.1
onemkl-license==2025.3.0
onnxruntime==1.23.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pefile==2024.8.26
pillow==12.0.0
pip==25.3
pluggy==1.6.0
protobuf==6.33.2
psutil==7.1.3
pyasn1==0.6.1
PyAutoGUI==0.9.54
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.10.6
pydantic_core==2.27.2
pydivert==2.1.0
PyGetWindow==0.0.9
Pygments==2.19.2
pyinstaller==6.18.0
pyinstaller-hooks-contrib==2025.10
PyJWT==2.10.1
PyMsgBox==2.0.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pyparsing==3.2.5
pyperclip==1.11.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.2
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-Levenshtein==0.27.3
pytweening==1.2.0
pytz==2025.2
pywin32==311
pywin32-ctypes==0.2.3
pywinauto==0.6.9
PyYAML==6.0.3
qiniu==7.17.0
RapidFuzz==3.14.3
requests==2.32.5
selenium==4.39.0
selenium-wire==5.1.0
setuptools==78.1.1
shellescape==3.8.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlacodegen==3.2.0
SQLAlchemy==2.0.45
sympy==1.14.0
tbb==2022.3.0
tcmlib==1.4.1
tenacity==9.1.2
trio==0.32.0
trio-websocket==0.12.2
typeguard==4.4.4
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
umf==1.0.2
undetected-chromedriver==3.5.5
urllib3==2.6.2
watchdog==6.0.0
websocket-client==1.9.0
websockets==14.2
Werkzeug==3.1.4
wheel==0.45.1
wsproto==1.3.2
xlsxwriter==3.2.9
zipp==3.23.0
zstandard==0.25.0
